refactor(wave7b-step2): mechanically split store helpers behind stable facade
